Migrate apperance plugin to new API
