+++ spyder/plugins/appearance/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+# (see spyder/__init__.py for details)
